+++ std/typeinfo/ti_AC.d
@@ -75,7 +75,7 @@ class TypeInfo_AC : TypeInfo
 	return c;
     }
 
-    int tsize()
+    size_t tsize()
     {
 	return (Object[]).sizeof;
     }
