+++ std/typeinfo/ti_Acreal.d
@@ -32,7 +32,7 @@ class TypeInfo_Ac : TypeInfo
 
     uint getHash(void *p)
     {	creal[] s = *cast(creal[]*)p;
-	uint len = s.length;
+	size_t len = s.length;
 	creal *str = s;
 	uint hash = 0;
 
@@ -55,7 +55,7 @@ class TypeInfo_Ac : TypeInfo
     {
 	creal[] s1 = *cast(creal[]*)p1;
 	creal[] s2 = *cast(creal[]*)p2;
-	uint len = s1.length;
+	size_t len = s1.length;
 
 	if (len != s2.length)
 	    return 0;
@@ -72,7 +72,7 @@ class TypeInfo_Ac : TypeInfo
     {
 	creal[] s1 = *cast(creal[]*)p1;
 	creal[] s2 = *cast(creal[]*)p2;
-	uint len = s1.length;
+	size_t len = s1.length;
 
 	if (s2.length < len)
 	    len = s2.length;
@@ -85,7 +85,7 @@ class TypeInfo_Ac : TypeInfo
 	return cast(int)s1.length - cast(int)s2.length;
     }
 
-    int tsize()
+    size_t tsize()
     {
 	return (creal[]).sizeof;
     }
