+++ std/typeinfo/ti_Ashort.d
@@ -9,7 +9,7 @@ class TypeInfo_As : TypeInfo
 
     uint getHash(void *p)
     {	short[] s = *cast(short[]*)p;
-	uint len = s.length;
+	size_t len = s.length;
 	short *str = s;
 	uint hash = 0;
 
@@ -50,11 +50,11 @@ class TypeInfo_As : TypeInfo
     {
 	short[] s1 = *cast(short[]*)p1;
 	short[] s2 = *cast(short[]*)p2;
-	uint len = s1.length;
+	size_t len = s1.length;
 
 	if (s2.length < len)
 	    len = s2.length;
-	for (uint u = 0; u < len; u++)
+	for (size_t u = 0; u < len; u++)
 	{
 	    int result = s1[u] - s2[u];
 	    if (result)
@@ -63,7 +63,7 @@ class TypeInfo_As : TypeInfo
 	return cast(int)s1.length - cast(int)s2.length;
     }
 
-    int tsize()
+    size_t tsize()
     {
 	return (short[]).sizeof;
     }
